Fix routeHighlighter content for cancel + delete tasks
Both of these endpoints require at least one query parameter, but it doesn't necessarily have to be a task_uid. And in fact, if you try to give just a task uid as a parameter, it will throw a deserialize error. Task uids must follow the uids query parameter, e.g. uids=5,6