Update the title Multi-index search in the api refs by removing index (for the new doc)
Using the term multi-index search implies that this route is meant to search on multiple indexes only. Which is not the case. This route is meant to perform multiple search on any index, may it be 4 times on the same index.

Also, the name of the route is multi-search and the feature is named multi search, it seems more consistent to keep that name.