Distance is the sum, not the max of distance values between words
In the example "Bruce Willis,Vin Diesel" there is not only a comma (8) but also a space (1) between Bruce and Vin, hence a distance of 9 instead of 8. Same applies to the subsequent examples in the following sentence. Otherwise it would be inconsistent with the next JSON example and the general concept of proximity outlined elsewhere in the documentation in which distances are plausibly characterized as additive.