fix: update download link for movies.json file
The Link of the `movies.json` file was not working in the rendered page. There I got an 404 pointing to https://meilisearch.com/assets/datasets/movies.json

It was only working from the source file on GitHub which points to https://github.com/meilisearch/documentation/blob/main/assets/datasets/movies.json

So I decided to update the link to point to https://www.meilisearch.com/movies.json which should work for both.